Build distinct versions for standalone and portal, bail out if emulated
Trying to detect at runtime whether or not the portal code is present
is too complex to do reliably, since this would need proper checksumming
of the code and the charset. It is much easier to simply build two
distinct versions for portal or standalone usage.

The Makefile now builds both wic64-update.prg and wic64-update-portal.prg.

Also use the new command WIC64_IS_HARDWARE and bail out early if
the WiC64 is emulated, exiting with a proper error message.